==> 1-Data Cleaning.ipynb <==
import requests
from bs4 import BeautifulSoup
import pickle


def url_to_transcript(url):
    '''Returns transcript data specifically from nytimes.com.'''
    page = requests.get(url).text
    soup = BeautifulSoup(page, "lxml")
    text = [p.text for p in soup.find(class_="meteredContent css-1r7ky0e").find_all('p')]
    print(url)
    return text

urls = ['https://www.nytimes.com/2020/06/19/opinion/capitol-confederacy-statues.html/',
        'https://www.nytimes.com/2020/06/17/opinion/george-floyd-arab-muslims-racism.html/',
        'https://www.nytimes.com/2020/06/14/opinion/george-floyd-psalms-bible.html/',
        'https://www.nytimes.com/2020/06/01/opinion/george-floyd-protest-police.html/',
        'https://www.nytimes.com/2020/05/29/opinion/Minneapolis-police-George-Floyd.html/',
        'https://www.nytimes.com/2020/06/08/opinion/george-floyd-protests-race.html/',
        'https://www.nytimes.com/2020/06/02/opinion/george-floyd-protests-first-amendment.html/',
        'https://www.nytimes.com/2020/06/06/opinion/sunday/george-floyd-structural-racism.html/']

editorials = ['1', '2', '3', '4', '5', '6', '7', '8']


# # Actually request transcripts (takes a few minutes to run)
transcripts = [url_to_transcript(u) for u in urls]

# # Pickle files for later use

# # Make a new directory to hold the text files
!mkdir transcripts

for i, c in enumerate(editorials):
     with open("transcripts/" + c + ".txt", "wb") as file:
         pickle.dump(transcripts[i], file)

# Load pickled files
data = {}
for i, c in enumerate(editorials):
    with open("transcripts/" + c + ".txt", "rb") as file:
        data[c] = pickle.load(file)

# Double check to make sure data has been loaded properly
data.keys()

# More checks
data['1'][:4]

# Let's take a look at our data again
next(iter(data.keys()))

# Notice that our dictionary is currently in key: editorial, value: list of text format
next(iter(data.values()))

# We are going to change this to key: editorial, value: string format
def combine_text(list_of_text):
    '''Takes a list of text and combines them into one large chunk of text.'''
    combined_text = ' '.join(list_of_text)
    return combined_text

# Combine it!
data_combined = {key: [combine_text(value)] for (key, value) in data.items()}

# We can either keep it in dictionary format or put it into a pandas dataframe
import pandas as pd
pd.set_option('max_colwidth',150)

data_df = pd.DataFrame.from_dict(data_combined).transpose()
data_df.columns = ['transcript']
data_df = data_df.sort_index()
data_df

# Let's take a look at the transcript for Ali Wong
data_df.transcript.loc['7']

# Apply a first round of text cleaning techniques
import re
import string

def clean_text_round1(text):
    '''Make text lowercase, remove text in square brackets, remove punctuation and remove words containing numbers.'''
    text = text.lower()
    text = re.sub('\[.*?\]', '', text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('\w*\d\w*', '', text)
    return text

round1 = lambda x: clean_text_round1(x)

# Let's take a look at the updated text
data_clean = pd.DataFrame(data_df.transcript.apply(round1))
data_clean

# Apply a second round of cleaning
def clean_text_round2(text):
    '''Get rid of some additional punctuation and non-sensical text that was missed the first time around.'''
    text = re.sub('[‘’“”…]', '', text)
    text = re.sub('\n', '', text)
    return text

round2 = lambda x: clean_text_round2(x)

# Let's take a look at the updated text
data_clean = pd.DataFrame(data_clean.transcript.apply(round2))
data_clean

# Let's take a look at our dataframe
data_df

# Let's add the editorias' full names as well
full_names = ['1', '2', '3', '4', '5', '6', '7', '8']

data_df['full_name'] = full_names
data_df

# Let's pickle it for later use
data_df.to_pickle("corpus.pkl")

# We are going to create a document-term matrix using CountVectorizer, and exclude common English stop words
from sklearn.feature_extraction.text import CountVectorizer

cv = CountVectorizer(stop_words='english')
data_cv = cv.fit_transform(data_clean.transcript)
data_dtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
data_dtm.index = data_clean.index
data_dtm

# Let's pickle it for later use
data_dtm.to_pickle("dtm.pkl")

# Let's also pickle the cleaned data (before we put it in document-term matrix format) and the CountVectorizer object
data_clean.to_pickle('data_clean.pkl')
pickle.dump(cv, open("cv.pkl", "wb"))
